+++ blastermaster/Animations/Animations.cpp
@@ -1,5 +1,5 @@
 #include "Animations.h"
-#include "Utils.h"
+#include "..\Utils\Utils.h"
 
 CAnimationSets* CAnimationSets::__instance = NULL;
