+++ blastermaster/Animations/Sprites.cpp
@@ -1,6 +1,6 @@
 #include "Sprites.h"
-#include "Game.h"
-#include "Utils.h"
+#include "..\Game\Game.h"
+#include "..\Utils\Utils.h"
 
 CSprite::CSprite(int id, int left, int top, int right, int bottom, LPDIRECT3DTEXTURE9 tex)
 {
