+++ blastermaster/Animations/Textures.cpp
@@ -3,8 +3,8 @@
 #include <d3d9.h>
 #include <d3dx9.h>
 
-#include "Utils.h"
-#include "Game.h"
+#include "..\Utils\Utils.h"
+#include "..\Game\Game.h"
 #include "textures.h"
 
 CTextures* CTextures::__instance = NULL;
