+++ blastermaster/Camera/Camera.cpp
