+++ blastermaster/Define/Define.cpp
