+++ blastermaster/Object/GameObject.cpp
@@ -1,10 +1,8 @@
 #include <algorithm>
 #include <d3dx9.h>
-#include "debug.h"
-#include "Game.h"
+#include "..\Utils\Utils.h"
+#include "..\Game\Game.h"
 #include "GameObject.h"
-#include "Animations.h"
-#include "Textures.h"
 
 CGameObject::CGameObject()
 {
