+++ blastermaster/Object/OutdoorEnemy/OutdoorEnemy.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "..\..\GameObject.h"
+#include "..\GameObject.h"
 
 enum OutdoorEnemyType {
 	Crabullus,
