+++ blastermaster/Utils/Utils.cpp
