+++ blastermaster/blastermaster.vcxproj
@@ -148,77 +148,77 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Animations.cpp" />
+    <ClCompile Include="Animations\Animations.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\BossCrabullus.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Plutonium.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Cannon.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Dasher.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Dome.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\OutdoorEnemy.cpp" />
-    <ClCompile Include="Camera.cpp" />
-    <ClCompile Include="Define.cpp" />
+    <ClCompile Include="Camera\Camera.cpp" />
+    <ClCompile Include="Define\Define.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Eyeball.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Floater.cpp" />
-    <ClCompile Include="Game.cpp" />
-    <ClCompile Include="GameObject.cpp" />
+    <ClCompile Include="Game\Game.cpp" />
+    <ClCompile Include="Object\GameObject.cpp" />
     <ClCompile Include="Object\IndoorEnemy\IndoorEnemy.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Insect.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Jumper.cpp" />
-    <ClCompile Include="Grid.cpp" />
+    <ClCompile Include="Grid\Grid.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Map.cpp" />
+    <ClCompile Include="Map\Map.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Mine.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Orb.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Ship.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Skull.cpp" />
     <ClCompile Include="Object\Player\Sophia.cpp" />
-    <ClCompile Include="Scene.cpp" />
-    <ClCompile Include="SceneGame.cpp" />
-    <ClCompile Include="Sprites.cpp" />
+    <ClCompile Include="Scene\Scene.cpp" />
+    <ClCompile Include="Scene\SceneGame.cpp" />
+    <ClCompile Include="Animations\Sprites.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Teleporter.cpp" />
-    <ClCompile Include="Textures.cpp" />
+    <ClCompile Include="Animations\Textures.cpp" />
     <ClCompile Include="Object\IndoorEnemy\UnderworldLord.cpp" />
     <ClCompile Include="Object\IndoorEnemy\Waver.cpp" />
     <ClCompile Include="Object\OutdoorEnemy\Worm.cpp" />
-    <ClCompile Include="Utils.cpp" />
+    <ClCompile Include="Utils\Utils.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Animations.h" />
-    <ClInclude Include="KeyEventHandler.h" />
+    <ClInclude Include="Animations\Animations.h" />
+    <ClInclude Include="Scene\KeyEventHandler.h" />
     <ClInclude Include="Object\OutdoorEnemy\BossCrabullus.h" />
     <ClInclude Include="Object\IndoorEnemy\Plutonium.h" />
     <ClInclude Include="Object\IndoorEnemy\Cannon.h" />
     <ClInclude Include="Object\IndoorEnemy\Dasher.h" />
     <ClInclude Include="Object\OutdoorEnemy\Dome.h" />
     <ClInclude Include="Object\OutdoorEnemy\OutdoorEnemy.h" />
-    <ClInclude Include="Camera.h" />
-    <ClInclude Include="Define.h" />
+    <ClInclude Include="Camera\Camera.h" />
+    <ClInclude Include="Define\Define.h" />
     <ClInclude Include="Object\IndoorEnemy\Eyeball.h" />
     <ClInclude Include="Object\OutdoorEnemy\Floater.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="GameObject.h" />
+    <ClInclude Include="Game\Game.h" />
+    <ClInclude Include="Object\GameObject.h" />
     <ClInclude Include="Object\IndoorEnemy\IndoorEnemy.h" />
     <ClInclude Include="Object\OutdoorEnemy\Insect.h" />
     <ClInclude Include="Object\OutdoorEnemy\Jumper.h" />
-    <ClInclude Include="Grid.h" />
-    <ClInclude Include="Map.h" />
+    <ClInclude Include="Grid\Grid.h" />
+    <ClInclude Include="Map\Map.h" />
     <ClInclude Include="Object\OutdoorEnemy\Mine.h" />
     <ClInclude Include="Object\OutdoorEnemy\Orb.h" />
     <ClInclude Include="Object\OutdoorEnemy\Ship.h" />
     <ClInclude Include="Object\OutdoorEnemy\Skull.h" />
     <ClInclude Include="Object\Player\Sophia.h" />
-    <ClInclude Include="Scene.h" />
-    <ClInclude Include="SceneGame.h" />
-    <ClInclude Include="Sprites.h" />
+    <ClInclude Include="Scene\Scene.h" />
+    <ClInclude Include="Scene\SceneGame.h" />
+    <ClInclude Include="Animations\Sprites.h" />
     <ClInclude Include="Object\IndoorEnemy\Teleporter.h" />
-    <ClInclude Include="Textures.h" />
+    <ClInclude Include="Animations\Textures.h" />
     <ClInclude Include="Object\IndoorEnemy\UnderworldLord.h" />
     <ClInclude Include="Object\IndoorEnemy\Waver.h" />
     <ClInclude Include="Object\OutdoorEnemy\Worm.h" />
-    <ClInclude Include="Utils.h" />
+    <ClInclude Include="Utils\Utils.h" />
   </ItemGroup>
   <ItemGroup>
-    <Text Include="map2.txt">
+    <Text Include="Map\map2.txt">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">false</ExcludedFromBuild>
     </Text>
   </ItemGroup>
